use actual public key as part of rng seed
Using the actual public key instead of the hash of the public key adds a bit more security as the actual public key is never transmitted. Only the hash of it. Move client checks out of socket timeouts. It might be possible for an attacker to send a steady stream of packets to prevent client checks for ever occuring. Rename/redocument constants to make them more clear what they do.